% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene.converter.R
\name{GeneConverter}
\alias{GeneConverter}
\title{Converting Gene Symbols or FlyBase IDs to a certain version Flybase version.}
\usage{
GeneConverter(x, version, verbose = F)
}
\arguments{
\item{x}{a vector of gene symbols.}

\item{version}{FlyBase ID version for the updated result. It should be either version numbers (e.g. 6.12) or FlyBase Release Numbers (e.g. FB2017_04).  The default is "current".}
}
\description{
The function takes Gene Symbols or FlyBase IDs (e.g. FBgn#######) as an input, and converts it into certain versions of IDs.
The function accesses the FlyBase FTP site, so requires internet-connection.
FlyBase IDs for genes that are split into multiple genes will be concatenated with two colons (::).  Genes that does not have matching IDs will be shown as "unknown".
}
\examples{
GeneConverter(x, version=6.12)
GeneConverter(x, version="FB2016_04")
}
\keyword{flybase}
