% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene.converter.R
\name{GetFlybaseAnnotation}
\alias{GetFlybaseAnnotation}
\title{Get Flybase annotation ID file given the version}
\usage{
GetFlybaseAnnotation(version)
}
\arguments{
\item{version}{FlyBase ID version.}

\item{x}{a vector of gene symbols.}
}
\description{
The function takes Flybase version as an input, and downlooad the the Flybase annotation ID file. 
Then if filter out non-melanogaster genes and copy information from annotation.ID to secondary.annotation.ID if latter is empty
}
\details{
The function accesses the FlyBase FTP site, so requires internet-connection.
}
\examples{
GeneConverter(x, version="FB2016_04")
}
\keyword{flybase}
